remove as u64

+++ program/src/deposit.rs
@@ -177,7 +177,7 @@ fn state_transistion(req: StateTransistionRequest) -> Result<(), FundError> {
             vault_acc_info.key,
             depositor_authority_acc_info.key,
             &[],
-            amount as u64,
+            amount,
         )?;
         program::invoke_signed(
             &deposit_instruction,
